scripts/movegen_diff/requirements.txt
progressbar2==3.30.2
python-chess==0.18.4
python-utils==2.1.0
six==1.10.0
